Implement Sisimai::RFC791, Sisimai::String.ipv4 has been moved to Sisimai::RFC791.find #319
